move DocumentElement
