+++ packages/flame/lib/src/text/block/document_element.dart
