+++ packages/flame/lib/src/text/styles/document_style.dart
@@ -1,4 +1,4 @@
-import 'package:flame/src/text/document_element.dart';
+import 'package:flame/src/text/block/document_element.dart';
 import 'package:flame/src/text/nodes.dart';
 import 'package:flame/src/text/styles/block_style.dart';
 import 'package:flutter/painting.dart';
